Alibaba OA
One pretty interesting question:
If you are given an array e.g. [113, 215, 221], return all path sum from root to leaf in the tree. The max depth of the tree is 4.
Description for the array:
for each individual element:
1st digit denotes the # of level: 1 -> root 
2nd digit denotes the position in the level, ranged from 1-8
3rd digit denotes the value of the node

So we have the following tree:
        3 
      /    \
   5        1

So the solution of the sample array is: (3+5) + (3+1) = 12